feat(Favorite*.*): 즐겨찾기 결과를 복수로 설정
- 잘못 된 응답 수정

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,11 +1,14 @@
 package nextstep.subway.favorite.application;
 
-import java.util.Optional;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import javax.persistence.NoResultException;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import nextstep.subway.favorite.dto.FavoriteResponse;
+import nextstep.subway.favorite.dto.FavoritesResponse;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.member.domain.Member;
@@ -14,6 +17,7 @@ import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
+
 
 @Service
 @Transactional
@@ -37,33 +41,43 @@ public class FavoriteService {
         Station source = stationRepository.findById(favoriteRequest.getSource()).orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_STATION));
         Station target = stationRepository.findById(favoriteRequest.getTarget()).orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_STATION));
 
-        Optional<Favorite> favoriteOptional = favoriteRepository.findByMember(member);
-        if (favoriteOptional.isPresent()) {
+        List<Favorite> favorites = favoriteRepository.findByMember(member);
+        boolean isPresent = favorites.stream()
+                .map(favorite -> favorite.getSource().equals(source) && favorite.getTarget().equals(target))
+                .findAny().isPresent();
+
+        if (isPresent) {
             throw new IllegalArgumentException(FAVORITE_ALREADY_ADDED);
         }
 
         favoriteRepository.save(new Favorite(member, source, target));
     }
 
-    public FavoriteResponse search(LoginMember loginMember) {
+    public FavoritesResponse search(LoginMember loginMember) {
         Member member = member(loginMember);
-        Favorite favorite = getFavorite(member);
+        List<Favorite> favorites = getFavorites(member);
 
-        return FavoriteResponse.of(favorite);
+        return FavoritesResponse.of(favorites.stream().map(Favorite::toResponse).collect(Collectors.toList()));
     }
 
     public void remove(LoginMember loginMember) {
         Member member = member(loginMember);
-        Favorite favorite = getFavorite(member);
+        List<Favorite> favorites = getFavorites(member);
 
-        favoriteRepository.delete(favorite);
+        if (favorites.isEmpty()) {
+            throw new NoResultException(NOT_FOUND_FAVORITE);
+        }
+
+        for (Favorite favorite : favorites) {
+            favoriteRepository.delete(favorite);
+        }
     }
 
     private Member member(LoginMember loginMember) {
-        return memberRepository.findById(loginMember.getId()).orElseThrow(RuntimeException::new);
+        return memberRepository.findById(loginMember.getId()).orElseThrow(IllegalArgumentException::new);
     }
 
-    private Favorite getFavorite(Member member) {
-        return favoriteRepository.findByMember(member).orElseThrow(() -> new RuntimeException(NOT_FOUND_FAVORITE));
+    private List<Favorite> getFavorites(Member member) {
+        return favoriteRepository.findByMember(member);
     }
 }
